Avoid a null pointer dereference of node pointer.
Use json objects on heap, as I now think is required by the library, as it
tries to free them.

Go back to doing reporting on a thread, because meshtastic thing didn't work,
and now seem to have enough memory.

Initialse JSON objects. Don't know if needed but doesn't hurt.